requirements.txt
openai>=1.68.0
openai-agents>=0.0.15
sendgrid>=6.11.0
python-dotenv>=1.0.1
pydantic>=2.0.0
